feat: add ListsTab, RequestsTab, and TasksTab components for class management
- Implemented ListsTab to display lists associated with a class.
- Created RequestsTab to handle student class permission requests with approve/reject functionality.
- Developed TasksTab to manage tasks, including viewing submissions and linking new tasks.
- Updated ClassPage to integrate new tabs and streamline class-related functionalities.
