+++ packages/vulpes/src/app/(class)/class/[ID]/page.tsx
@@ -1,119 +1,30 @@
 "use client";
 
 import { IClass } from "@/@types/Class";
-import { ISubmission } from "@/@types/Submission";
-import { ITask } from "@/@types/Task";
+
 import AppNavBar from "@/components/AppNavBar";
 import AuthGuard from "@/components/AuthGuard";
 import { useAuth } from "@/providers/AuthProvider";
 import API from "@/services/API";
-import { Editor } from "@monaco-editor/react";
-import ArrowBackIcon from "@mui/icons-material/ArrowBack";
-import CheckBoxIcon from "@mui/icons-material/CheckBox";
-import CheckBoxOutlineBlankIcon from "@mui/icons-material/CheckBoxOutlineBlank";
+
 import {
   Alert,
-  Autocomplete,
   Box,
-  Button,
-  Card,
-  CardContent,
-  Checkbox,
   CircularProgress,
   Container,
-  Divider,
-  Modal,
-  Paper,
-  Stack,
   Tab,
-  Table,
-  TableBody,
-  TableCell,
-  TableContainer,
-  TableHead,
-  TableRow,
   Tabs,
-  TextField,
   Typography,
 } from "@mui/material";
-import { AnimatePresence, motion } from "framer-motion";
-import Link from "next/link";
 import { useParams } from "next/navigation";
-import { FormEvent, useEffect, useMemo, useState } from "react";
-import { toast } from "react-toastify";
-import { registerPortugolLanguage } from "../../../../../libs/monaco-config";
-import MDEditor from "@uiw/react-md-editor";
+import { useEffect, useMemo, useState } from "react";
 
-interface IClassRequest {
-  classId: string;
-  studentId: string;
-  message: string | null;
-  createdAt: string;
-  student?: {
-    userId: string;
-    name: string;
-    email: string;
-  };
-}
-
-interface IGetClassRequestsResponse {
-  requests: IClassRequest[];
-  total: number;
-  page: number;
-  limit: number;
-  totalPages: number;
-}
-
-interface IListItem {
-  listId: string;
-  classId: string;
-  title: string;
-  deadline: string;
-  submissionLimit: number;
-}
-
-interface IGetListsResponse {
-  lists: IListItem[];
-  total: number;
-  page: number;
-  limit: number;
-  totalPages: number;
-}
-
-interface ITaskItem {
-  taskId: string;
-  title: string;
-  description: string;
-  isPublic: boolean;
-  isVisible: boolean;
-  creatorId: string;
-  creator?: {
-    userId: string;
-    name: string;
-  };
-}
-
-interface IGetTasksResponse {
-  tasks: ITaskItem[];
-  total: number;
-  page: number;
-  limit: number;
-  totalPages: number;
-}
-
-const checkboxIcon = <CheckBoxOutlineBlankIcon fontSize="small" />;
-const checkboxCheckedIcon = <CheckBoxIcon fontSize="small" />;
+import RequestsTab from "./components/RequestTab";
+import CreateListTab from "./components/CreateListTab";
+import ListsTab from "./components/ListsTab";
+import TasksTab from "./components/TaskTab";
 
 export default function ClassPage() {
-  return (
-    <AuthGuard requiredRoles={["STUDENT", "PROFESSOR", "ADMIN"]}>
-      <AppNavBar />
-      <ClassContent />
-    </AuthGuard>
-  );
-}
-
-function ClassContent() {
   const { ID } = useParams();
   const { user } = useAuth();
   const classId = Array.isArray(ID) ? ID[0] : ID;
@@ -189,673 +100,47 @@ function ClassContent() {
   const activeTab = tabs[tabIndex]?.key ?? "lists";
 
   return (
-    <Container maxWidth="lg" sx={{ py: 4 }}>
-      <Box sx={{ mb: 3 }}>
-        <Typography variant="h4" component="h1">
-          {classData.name}
-        </Typography>
-        <Typography variant="body2" color="text.secondary">
-          Código: {classData.code} • Professor: {classData.professor.name}
-        </Typography>
-      </Box>
+    <AuthGuard requiredRoles={["STUDENT", "PROFESSOR", "ADMIN"]}>
+      <AppNavBar />
+      <Container maxWidth="lg" sx={{ py: 4 }}>
+        <Box sx={{ mb: 3 }}>
+          <Typography variant="h4" component="h1">
+            {classData.name}
+          </Typography>
+          <Typography variant="body2" color="text.secondary">
+            Código: {classData.code} • Professor: {classData.professor.name}
+          </Typography>
+        </Box>
 
-      <Tabs value={tabIndex} onChange={(_, v) => setTabIndex(v)} sx={{ mb: 3 }}>
-        {tabs.map((tab) => (
-          <Tab key={tab.key} label={tab.label} />
-        ))}
-      </Tabs>
-
-      {activeTab === "tasks" && (
-        <TasksTab
-          classId={classData.classId}
-          isProfessorOrAdmin={isProfessorOrAdmin}
-        />
-      )}
-      {activeTab === "requests" && <RequestsTab classId={classData.classId} />}
-      {activeTab === "create" && (
-        <CreateListTab
-          classId={classData.classId}
-          onCreated={() =>
-            setTabIndex(tabs.findIndex((t) => t.key === "lists"))
-          }
-        />
-      )}
-      {activeTab === "lists" && <ListsTab classId={classData.classId} />}
-    </Container>
-  );
-}
-
-interface ITaskTabProps {
-  classId: string;
-  isProfessorOrAdmin: boolean;
-}
-
-function TasksTab({ classId, isProfessorOrAdmin }: ITaskTabProps) {
-  const [tasks, setTasks] = useState<ITask[]>([]);
-  const [loading, setLoading] = useState(true);
-  const [showSubmissions, setShowSubmissions] = useState(false);
-  const [selectedTaskId, setSelectedTaskId] = useState<string | null>(null);
-  const [isModalOpen, setIsModalOpen] = useState(false);
-
-  const handleOpenModal = (taskId: string) => {
-    setIsModalOpen(true);
-  };
-
-  const handleCloseModal = () => {
-    setIsModalOpen(false);
-  };
-
-  const fetchTasks = async () => {
-    setLoading(true);
-    try {
-      const response = await API.get(`/class-task/class/${classId}`);
-      const filteredTasks = response.data.classTasks.map(
-        ({ task }: { task: ITask }) => task,
-      );
-
-      setTasks(filteredTasks);
-    } catch (error) {
-      console.error("Failed to fetch tasks:", error);
-      toast.error("Erro ao carregar tarefas.");
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  useEffect(() => {
-    fetchTasks();
-  }, [classId]);
-
-  function SubmissionsContent({
-    onBack,
-    taskId,
-  }: {
-    onBack: () => void;
-    taskId: string;
-  }) {
-    const [submissions, setSubmissions] = useState<ISubmission[]>([]);
-    const [selectedSubmission, setSelectedSubmission] =
-      useState<ISubmission | null>(null);
-
-    const fetchSubmissions = async () => {
-      const submissions = await API.get(`/submission/task/${taskId}`);
-      setSubmissions(submissions.data);
-    };
-
-    useEffect(() => {
-      fetchSubmissions();
-
-      return () => setSubmissions([]);
-    }, []);
-
-    const handleEditorDidMount = (editorInstance: any, monacoInstance: any) => {
-      registerPortugolLanguage(monacoInstance);
-      monacoInstance.editor.setTheme("vs-dark");
-    };
-
-    return (
-      <Box sx={{ display: "flex", flexDirection: "column", gap: 2 }}>
-        <Button
-          startIcon={<ArrowBackIcon />}
-          onClick={onBack}
-          variant="outlined"
-          sx={{ maxWidth: 128 }}
+        <Tabs
+          value={tabIndex}
+          onChange={(_, v) => setTabIndex(v)}
+          sx={{ mb: 3 }}
         >
-          Voltar
-        </Button>
-        <AnimatePresence mode="wait" initial={false}>
-          {selectedSubmission && (
-            <motion.div
-              key="sbumission-review"
-              initial={{ y: -100, opacity: 0 }}
-              animate={{ y: 0, opacity: 1 }}
-              exit={{ x: -100, opacity: 0 }}
-              transition={{ duration: 0.3 }}
-            >
-              <Divider sx={{ my: 2 }} />
-              <Stack direction="row" gap={1} sx={{ height: 300 }}>
-                <Editor
-                  height="100%"
-                  theme="vs-dark"
-                  language="portugol"
-                  onMount={handleEditorDidMount}
-                  value={selectedSubmission ? selectedSubmission.code : ""}
-                />
-                <MDEditor height="100%" style={{ width: "100%" }} />
-              </Stack>
-              <Stack
-                direction="row"
-                gap={1}
-                alignItems="center"
-                justifyContent="flex-end"
-                sx={{ my: 2 }}
-              >
-                <Button onClick={() => setSelectedSubmission(null)}>
-                  Cancelar
-                </Button>
-                <Button variant="contained">Enviar Feedback</Button>
-              </Stack>
-              <Divider />
-            </motion.div>
-          )}
-        </AnimatePresence>
+          {tabs.map((tab) => (
+            <Tab key={tab.key} label={tab.label} />
+          ))}
+        </Tabs>
 
-        <TableContainer component={Paper}>
-          <Table>
-            <TableHead>
-              <TableRow>
-                <TableCell width={320}>Nome</TableCell>
-                <TableCell>Email</TableCell>
-                <TableCell width={130} align="center">
-                  Resultado
-                </TableCell>
-                <TableCell width={130} align="center">
-                  Código
-                </TableCell>
-                <TableCell width={250}>Data</TableCell>
-              </TableRow>
-            </TableHead>
-            <TableBody>
-              {submissions.map((submission) => (
-                <TableRow key={submission.submissionId}>
-                  <TableCell>{submission.student.name}</TableCell>
-                  <TableCell>{submission.student.email}</TableCell>
-                  <TableCell align="center">
-                    {submission.isCorrect ? (
-                      <Typography color="success.main">Correto</Typography>
-                    ) : (
-                      <Typography color="error.main">Incorreto</Typography>
-                    )}
-                  </TableCell>
-                  <TableCell align="center">
-                    <Button onClick={() => setSelectedSubmission(submission)}>
-                      Analisar
-                    </Button>
-                  </TableCell>
-                  <TableCell>
-                    {new Date(submission.submittedAt).toLocaleString()}
-                  </TableCell>
-                </TableRow>
-              ))}
-            </TableBody>
-          </Table>
-        </TableContainer>
-      </Box>
-    );
-  }
-
-  return (
-    <Box sx={{ position: "relative", overflowX: "hidden", minHeight: 400 }}>
-      <AnimatePresence mode="wait" initial={false}>
-        {!showSubmissions ? (
-          <motion.div
-            key="list"
-            initial={{ x: -100, opacity: 0 }}
-            animate={{ x: 0, opacity: 1 }}
-            exit={{ x: -100, opacity: 0 }}
-            transition={{ duration: 0.3 }}
-          >
-            <Box sx={{ display: "flex", flexDirection: "column", gap: 2 }}>
-              {isProfessorOrAdmin && (
-                <Button onClick={handleOpenModal}>
-                  Víncular Novas Tarefas
-                </Button>
-              )}
-
-              {loading ? (
-                <CircularProgress />
-              ) : (
-                <TableContainer component={Paper}>
-                  <Table sx={{ minWidth: 650 }}>
-                    <TableHead>
-                      <TableRow>
-                        <TableCell width={120}></TableCell>
-                        <TableCell width={40}>Envios</TableCell>
-                        <TableCell>Título</TableCell>
-                        <TableCell align="center" width={100}>
-                          Dificuldade
-                        </TableCell>
-                        <TableCell align="center" width={150}>
-                          Data de Criação
-                        </TableCell>
-                      </TableRow>
-                    </TableHead>
-                    <TableBody>
-                      {tasks.map((task) => (
-                        <TableRow key={task.taskId}>
-                          <TableCell>
-                            <Link
-                              href={`/task/${task.taskId}`}
-                              style={{ textDecoration: "none" }}
-                            >
-                              <Typography variant="body2" color="primary">
-                                Ver Detalhes
-                              </Typography>
-                            </Link>
-                          </TableCell>
-                          <TableCell align="center">
-                            <Box
-                              onClick={() => {
-                                setShowSubmissions(true);
-                                setSelectedTaskId(task.taskId);
-                              }}
-                              sx={{
-                                cursor: "pointer",
-                                textDecoration: "underline",
-                                color: "primary.main",
-                                "&:hover": { fontWeight: "bold" },
-                              }}
-                            >
-                              Ver
-                            </Box>
-                          </TableCell>
-                          <TableCell component="th" scope="row">
-                            {task.title}
-                          </TableCell>
-                          <TableCell align="center">Fácil</TableCell>
-                          <TableCell align="center">
-                            {new Date(task.createdAt).toLocaleDateString()}
-                          </TableCell>
-                        </TableRow>
-                      ))}
-                    </TableBody>
-                  </Table>
-                </TableContainer>
-              )}
-              <Modal open={isModalOpen} onClose={handleCloseModal}>
-                <Box
-                  sx={{
-                    position: "absolute" as "absolute",
-                    top: "50%",
-                    left: "50%",
-                    transform: "translate(-50%, -50%)",
-                    width: { xs: "90%", sm: 500 },
-                    bgcolor: "background.paper",
-                    boxShadow: 24,
-                    p: 4,
-                  }}
-                >
-                  <Typography variant="h6" component="h2" sx={{ mb: 2 }}>
-                    Víncular Novas Tarefas
-                  </Typography>
-                  <Typography variant="body1">
-                    Esta funcionalidade ainda está em desenvolvimento.
-                  </Typography>
-                </Box>
-              </Modal>
-            </Box>
-          </motion.div>
-        ) : (
-          <motion.div
-            key="submissions"
-            initial={{ x: 100, opacity: 0 }}
-            animate={{ x: 0, opacity: 1 }}
-            exit={{ x: 100, opacity: 0 }}
-            transition={{ duration: 0.3 }}
-          >
-            <SubmissionsContent
-              onBack={() => setShowSubmissions(false)}
-              taskId={selectedTaskId!}
-            />
-          </motion.div>
-        )}
-      </AnimatePresence>
-    </Box>
-  );
-}
-
-function RequestsTab({ classId }: { classId: string }) {
-  const [requests, setRequests] = useState<IClassRequest[]>([]);
-  const [loading, setLoading] = useState(true);
-
-  const fetchRequests = async () => {
-    setLoading(true);
-    try {
-      const response = await API.get<IGetClassRequestsResponse>(
-        `/student-class-permission-request/class/${classId}`,
-        { params: { page: 1, limit: 50 } },
-      );
-      setRequests(response.data.requests);
-    } catch (error) {
-      console.error("Failed to fetch requests:", error);
-      toast.error("Erro ao carregar solicitações.");
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  useEffect(() => {
-    fetchRequests();
-  }, [classId]);
-
-  const handleApprove = async (studentId: string) => {
-    try {
-      await API.patch(
-        `/student-class-permission-request/${classId}/${studentId}/approve`,
-      );
-      toast.success("Solicitação aprovada.");
-      fetchRequests();
-    } catch (error) {
-      toast.error("Erro ao aprovar solicitação.");
-    }
-  };
-
-  const handleReject = async (studentId: string) => {
-    try {
-      await API.patch(
-        `/student-class-permission-request/${classId}/${studentId}/reject`,
-      );
-      toast.success("Solicitação rejeitada.");
-      fetchRequests();
-    } catch (error) {
-      toast.error("Erro ao rejeitar solicitação.");
-    }
-  };
-
-  if (loading) {
-    return (
-      <Box sx={{ display: "flex", justifyContent: "center", py: 8 }}>
-        <CircularProgress />
-      </Box>
-    );
-  }
-
-  if (requests.length === 0) {
-    return (
-      <Box sx={{ textAlign: "center", py: 8 }}>
-        <Typography color="text.secondary">
-          Nenhuma solicitação pendente.
-        </Typography>
-      </Box>
-    );
-  }
-
-  return (
-    <Box sx={{ display: "flex", flexDirection: "column", gap: 2 }}>
-      {requests.map((request) => (
-        <Card key={`${request.classId}-${request.studentId}`}>
-          <CardContent
-            sx={{
-              display: "flex",
-              justifyContent: "space-between",
-              alignItems: "center",
-              flexWrap: "wrap",
-              gap: 2,
-            }}
-          >
-            <Box>
-              <Typography variant="h6">
-                {request.student?.name ?? "Aluno"}
-              </Typography>
-              <Typography variant="body2" color="text.secondary">
-                {request.student?.email ?? ""}
-              </Typography>
-              {request.message && (
-                <Typography variant="body2" sx={{ mt: 1 }}>
-                  Mensagem: {request.message}
-                </Typography>
-              )}
-              <Typography variant="caption" color="text.secondary">
-                Enviada em:{" "}
-                {new Date(request.createdAt).toLocaleDateString("pt-BR")}
-              </Typography>
-            </Box>
-            <Box sx={{ display: "flex", gap: 1 }}>
-              <Button
-                variant="outlined"
-                color="error"
-                onClick={() => handleReject(request.studentId)}
-              >
-                Rejeitar
-              </Button>
-              <Button
-                variant="contained"
-                onClick={() => handleApprove(request.studentId)}
-              >
-                Aprovar
-              </Button>
-            </Box>
-          </CardContent>
-        </Card>
-      ))}
-    </Box>
-  );
-}
-
-function CreateListTab({
-  classId,
-  onCreated,
-}: {
-  classId: string;
-  onCreated: () => void;
-}) {
-  const { user } = useAuth();
-  const [title, setTitle] = useState("");
-  const [deadline, setDeadline] = useState("");
-  const [submissionLimit, setSubmissionLimit] = useState(1);
-  const [submitting, setSubmitting] = useState(false);
-  const [tasks, setTasks] = useState<ITaskItem[]>([]);
-  const [selectedTasks, setSelectedTasks] = useState<ITaskItem[]>([]);
-  const [loadingTasks, setLoadingTasks] = useState(true);
-
-  // Filter tasks: only show tasks that are (visible AND public) OR owned by the user
-  const availableTasks = useMemo(() => {
-    return tasks.filter((task) => {
-      const isVisibleAndPublic = task.isVisible && task.isPublic;
-      const isOwner = task.creatorId === user?.userId;
-      return isVisibleAndPublic || isOwner;
-    });
-  }, [tasks, user?.userId]);
-
-  useEffect(() => {
-    const fetchTasks = async () => {
-      setLoadingTasks(true);
-      try {
-        const response = await API.get<IGetTasksResponse>("/task", {
-          params: { page: 1, limit: 100 },
-        });
-        setTasks(response.data.tasks);
-      } catch (error) {
-        console.error("Failed to fetch tasks:", error);
-        toast.error("Erro ao carregar tarefas disponíveis.");
-      } finally {
-        setLoadingTasks(false);
-      }
-    };
-
-    fetchTasks();
-  }, []);
-
-  const handleSubmit = async (event: FormEvent) => {
-    event.preventDefault();
-
-    if (!title.trim()) {
-      toast.error("Informe o nome da lista.");
-      return;
-    }
-
-    if (!deadline) {
-      toast.error("Informe a data limite.");
-      return;
-    }
-
-    const parsedDate = new Date(deadline);
-    if (Number.isNaN(parsedDate.getTime())) {
-      toast.error("Data limite inválida.");
-      return;
-    }
-
-    setSubmitting(true);
-    try {
-      await API.post("/list", {
-        classId,
-        title: title.trim(),
-        deadline: parsedDate.toISOString(),
-        submissionLimit: submissionLimit || 1,
-        taskIds: selectedTasks.map((t) => t.taskId),
-      });
-      const taskCount = selectedTasks.length;
-      toast.success(
-        taskCount > 0
-          ? `Lista criada com ${taskCount} tarefa(s)!`
-          : "Lista criada com sucesso!",
-      );
-      setTitle("");
-      setDeadline("");
-      setSubmissionLimit(1);
-      setSelectedTasks([]);
-      onCreated();
-    } catch (error) {
-      toast.error("Erro ao criar lista.");
-    } finally {
-      setSubmitting(false);
-    }
-  };
-
-  return (
-    <Box component="form" onSubmit={handleSubmit} sx={{ maxWidth: 600 }}>
-      <TextField
-        fullWidth
-        label="Nome da lista"
-        value={title}
-        onChange={(e) => setTitle(e.target.value)}
-        sx={{ mb: 2 }}
-      />
-      <TextField
-        fullWidth
-        label="Data limite"
-        type="datetime-local"
-        value={deadline}
-        onChange={(e) => setDeadline(e.target.value)}
-        sx={{ mb: 2 }}
-        InputLabelProps={{ shrink: true }}
-      />
-      <TextField
-        fullWidth
-        label="Limite de envios"
-        type="number"
-        value={submissionLimit}
-        onChange={(e) => setSubmissionLimit(Number(e.target.value))}
-        inputProps={{ min: 1 }}
-        sx={{ mb: 2 }}
-      />
-      <Autocomplete
-        multiple
-        options={availableTasks}
-        disableCloseOnSelect
-        loading={loadingTasks}
-        value={selectedTasks}
-        onChange={(_, newValue) => setSelectedTasks(newValue)}
-        getOptionLabel={(option) => option.title}
-        isOptionEqualToValue={(option, value) => option.taskId === value.taskId}
-        renderOption={(props, option, { selected }) => {
-          const { key, ...otherProps } = props;
-          return (
-            <li key={option.taskId} {...otherProps}>
-              <Checkbox
-                icon={checkboxIcon}
-                checkedIcon={checkboxCheckedIcon}
-                style={{ marginRight: 8 }}
-                checked={selected}
-              />
-              <Box>
-                <Typography variant="body1">{option.title}</Typography>
-                <Typography variant="caption" color="text.secondary">
-                  {option.creator?.name ?? "Autor desconhecido"}
-                  {option.isPublic ? " • Pública" : ""}
-                </Typography>
-              </Box>
-            </li>
-          );
-        }}
-        renderInput={(params) => (
-          <TextField
-            {...params}
-            label="Tarefas (opcional)"
-            placeholder={
-              selectedTasks.length === 0 ? "Selecione tarefas..." : ""
-            }
-            InputProps={{
-              ...params.InputProps,
-              endAdornment: (
-                <>
-                  {loadingTasks ? (
-                    <CircularProgress color="inherit" size={20} />
-                  ) : null}
-                  {params.InputProps.endAdornment}
-                </>
-              ),
-            }}
+        {activeTab === "tasks" && (
+          <TasksTab
+            classId={classData.classId}
+            isProfessorOrAdmin={isProfessorOrAdmin}
           />
         )}
-        sx={{ mb: 3 }}
-      />
-      <Button type="submit" variant="contained" disabled={submitting}>
-        {submitting ? <CircularProgress size={24} /> : "Criar Lista"}
-      </Button>
-    </Box>
-  );
-}
-
-function ListsTab({ classId }: { classId: string }) {
-  const [lists, setLists] = useState<IListItem[]>([]);
-  const [loading, setLoading] = useState(true);
-
-  const fetchLists = async () => {
-    setLoading(true);
-    try {
-      const response = await API.get<IGetListsResponse>(
-        `/list/class/${classId}`,
-        { params: { page: 1, limit: 50 } },
-      );
-      setLists(response.data.lists);
-    } catch (error) {
-      console.error("Failed to fetch lists:", error);
-      toast.error("Erro ao carregar listas.");
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  useEffect(() => {
-    fetchLists();
-  }, [classId]);
-
-  if (loading) {
-    return (
-      <Box sx={{ display: "flex", justifyContent: "center", py: 8 }}>
-        <CircularProgress />
-      </Box>
-    );
-  }
-
-  if (lists.length === 0) {
-    return (
-      <Box sx={{ textAlign: "center", py: 8 }}>
-        <Typography color="text.secondary">
-          Nenhuma lista cadastrada.
-        </Typography>
-      </Box>
-    );
-  }
-
-  return (
-    <Box
-      sx={{
-        display: "grid",
-        gridTemplateColumns: { xs: "1fr", sm: "1fr 1fr", md: "1fr 1fr 1fr" },
-        gap: 2,
-      }}
-    >
-      {lists.map((list) => (
-        <Card key={list.listId}>
-          <CardContent>
-            <Typography variant="h6" noWrap>
-              {list.title}
-            </Typography>
-          </CardContent>
-        </Card>
-      ))}
-    </Box>
+        {activeTab === "requests" && (
+          <RequestsTab classId={classData.classId} />
+        )}
+        {activeTab === "create" && (
+          <CreateListTab
+            classId={classData.classId}
+            onCreated={() =>
+              setTabIndex(tabs.findIndex((t) => t.key === "lists"))
+            }
+          />
+        )}
+        {activeTab === "lists" && <ListsTab classId={classData.classId} />}
+      </Container>
+    </AuthGuard>
   );
 }
